Move UI test into a separate spec file and fix settings save button selector
