+++ plugins/CoreAdminHome/tests/UI/expected-screenshots/CustomLogoUpload_branding_settings_saved.png
