+++ plugins/CoreAdminHome/tests/UI/expected-screenshots/CustomLogoUpload_custom_favicon_uploaded.png
