+++ plugins/CoreAdminHome/tests/UI/expected-screenshots/CustomLogoUpload_custom_logo_uploaded.png
