+++ plugins/CoreAdminHome/tests/UI/expected-screenshots/CustomLogoUpload_disable_logo_upload.png
