+++ plugins/CoreAdminHome/tests/UI/expected-screenshots/CustomLogoUpload_enable_logo_upload.png
