--- plugins/CoreAdminHome/tests/UI/expected-screenshots/CustomLogo_Upload_settings_saved.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:aee9415e0e3052c1cebf754a0303ff5f00be39f1386dcd9c67f2c1019b02e1c4
-size 1318140
